refactor: viewName을 반환하는 컨트롤러와 객체를 반환하는 컨트롤러 분리
